backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
Pillow==10.0.1
python-decouple==3.8
dj-database-url==2.1.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
gunicorn==21.2.0
whitenoise==6.6.0
coverage==7.3.2
pytest-django==4.7.0
factory-boy==3.3.0
psycopg2-binary==2.9.9